requirements.txt
langchain-community==0.3.5
PyPDF2==3.0.1
pypdf==5.1.0
sentence-transformers==3.2.1
faiss-cpu==1.9.0
langchain-groq==0.2.1
pyglet==2.0.18
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
sounddevice==0.4.6
pyinstaller~=6.11.0
scipy==1.14.1
noisereduce==3.0.3
elevenlabs==1.13.3
streamlit==1.42.0
audio-recorder-streamlit==0.0.10
faster-whisper==1.1.1

python-dotenv~=1.0.1
websockets~=14.2
langchain~=0.3.18
torch~=2.6.0
transformers~=4.48.3
groq~=0.18.0
pydub~=0.25.1
opencv-python==4.11.0.86
